+++ open_audio.py
@@ -2,6 +2,7 @@ from pydub import AudioSegment
 import numpy as np
 import scipy.signal as ss
 import matplotlib.pyplot as plt 
+from utils import *
 
 fname = "Heart Of Glass.wav"
 format = 'wav'
@@ -22,24 +23,16 @@ right_ch = samples[1::2]
 left_ch = left_ch[:(fs*60)] 
 t = np.arange(0,60,1/fs)
 
-left_ch = left_ch/np.max(left_ch)
+left_ch = left_ch/np.max(left_ch) #normalize to avoid overflow in processing
 
-def plot_est_spectrum(x,fs,nperseg=65536):
-    f, Pxx_left = ss.welch(left_ch,fs,nperseg=nperseg)
-    plt.semilogy(f,Pxx_left)
-    plt.grid()
-    # plt.plot(f, Pxx_left)
-    plt.show()
-    
-def onset_func(x):
-    E = np.power(x,2)
-    dE_pos = np.diff(E,append=0)
-    dE_pos[dE_pos<0] = 0
-    O_t = dE_pos-np.mean(dE_pos)
-    return O_t
-
+plt.figure(1)
+plt.title("Welch PSD Estimation")
 plot_est_spectrum(left_ch,fs,4096)
+
+
+plt.figure(2)
 O_t = onset_func(left_ch)
 plt.plot(t,O_t)
+plt.title('Onset Function')
 plt.show()
 a=5
